refactor: useSafeActivityParams 파일화
